extracted ItemDndContext Component
